third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-dpr-zoom-expected.txt

PASS document.getElementById("foo").clientWidth==200 is true
This test passes if the image below is a 200px wide green square when the deviceScaleFactor is 1, since the user has zoomed in.
